daniel

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -3,7 +3,7 @@ package com.techelevator.tenmo.dao;
 import com.techelevator.tenmo.model.Account;
 
 public interface AccountDao {
-    Account getAccount(int id);
+    Account getAccount(Long id);
     void updateBalance(Account account);
 }
